bitwise

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -8,7 +8,7 @@
  */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned int d;
+	unsigned long int d;
 
 	if (index > (sizeof(unsigned long int) * 8 - 1))
 		return (-1);
